Move error messages to React
