+++ app/webpacker/components/RegistrationsV2/Register/RegistrationNotAllowedMessage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { List, Message } from 'semantic-ui-react';
+
+export default function RegistrationNotAllowedMessage({ reasons }) {
+  return (
+    <Message negative>
+      <List>
+        {reasons.map((reason) => (
+          <List.Item key={reason}>{reason}</List.Item>
+        ))}
+      </List>
+    </Message>
+  );
+}
